# Add PCH builds via 'SFML_ENABLE_PCH' CMake option