## Description

Enable PCH via CMake. Achieves roughly a 2x fresh compilation speedup on my machine, using MSYS2/MinGW. Looking for feedback on the direction and approach.

---

<details>
<summary>Machine Specs</summary>

- GIGABYTE Z390 AORUS MASTER-CF
- i9-9900k (stock settings)
- Corsair CMK16GX4M2B3000C15, 4x8GB sticks (XMP enabled)

</details>

---

<details>
<summary>Benchmark</summary>

Methodology:

```
cd build
rm -Rf src test examples

cmake .. -DCMAKE_BUILD_TYPE=Debug -GNinja -DSFML_BUILD_EXAMPLES=1 \ 
    -DCMAKE_EXPORT_COMPILE_COMMANDS=1 -DCMAKE_CXX_COMPILER=clang++ \
    -DCMAKE_CXX_FLAGS="-ftime-trace"

time ninja
```

Results without this PR:
- `ninja  0.00s user 0.01s system 0% cpu 5.599 total`

Results with this PR:
- `ninja  0.00s user 0.01s system 0% cpu 9.586 total`

</details>

---

This PR is related to the issue #1893.

## Tasks

* [ ] Tested on Linux
* [x] Tested on Windows
* [ ] Tested on macOS
* [ ] Tested on iOS
* [ ] Tested on Android

## How to test this PR?

Not sure yet.