@vittorioromeo 2021-12-19
@binary1248: any idea why CI is unhappy with the changes? The GCC error is particularly weird...