@binary1248 2021-12-20
I don't have any experience using PCH, definitely not with CMake.