@sam20908 2021-12-21
Did you ensure that the targets reusing the PCH have the same compiler definitions, options, and flags?

Quoted from CMake documentation:
> The REUSE_FROM form requires the same set of compiler options, compiler flags and compiler definitions for both <target> and <other_target>. Some compilers (e.g. GCC) may issue a warning if the precompiled header file cannot be used (-Winvalid-pch).