@vittorioromeo 2021-12-21
> Did you ensure that the targets reusing the PCH have the same compiler definitions, options, and flags?
> 
> Quoted from CMake documentation:
> 
> > The REUSE_FROM form requires the same set of compiler options, compiler flags and compiler definitions for both  and <other_target>. Some compilers (e.g. GCC) may issue a warning if the precompiled header file cannot be used (-Winvalid-pch).

There are no targets reusing the PCH, I've commented all of that out... The CI still fails with cryptic errors even when enabling PCH just for `sfml-system`.