@vittorioromeo 2022-02-28
Hah, yes! That was the issue. Removing `-Wpedantic` from `CompilerWarnings.cmake` makes PCH work! I don't understand why though, as we do call `set_file_warnings` on the PCH header. Does anybody want to investigate this further? @ChrisThrasher