@ChrisThrasher 2022-02-28
Precompiled headers require CMake 3.16. I see this PR doesn't change the minimum CMake version requirement. Is it possible that some jobs are failing due to old CMake versions?