@vittorioromeo 2022-10-01
At this point in time, PCH files seem to work only for Linux and MinGW, and only when SFML is built as a static library. For Visual Studio, the issue is:

```
CMake Error at cmake/Macros.cmake:[69](https://github.com/SFML/SFML/actions/runs/3165984825/jobs/5155401526#step:8:70) (add_library):
  PRECOMPILE_HEADERS_REUSE_FROM property is set on target ("sfml-window").
  Reusable precompile headers requires the COMPILE_PDB_NAME property to have
  the value "sfml-system"

Call Stack (most recent call first):
  src/SFML/Window/CMakeLists.txt:255 (sfml_add_library)
```

So to make it work, `COMPILE_PDB_NAME` has to be the same as the PCH target we are reusing the PCH from.

Regarding shared library builds: I cannot make PCH reuse work because of differing macro defines for `SFML_XXX_EXPORTS`. All targets must have the same macros. I can try having a per-library PCH, but I'm not sure if it's worth it.