@vittorioromeo 2022-10-01
I have no idea what the MacOS issue is about:

```
CMake Error in src/SFML/Window/CMakeLists.txt:
  Unable to resolve full path of PCH-header
  '/Users/runner/work/SFML/SFML/build/src/SFML/System/CMakeFiles/sfml-system.dir/cmake_pch.objc.h'
  assigned to target sfml-window, although its path is supposed to be known!
```