@vittorioromeo 2022-10-01
@ChrisThrasher @eXpl0it3r @binary1248 @Bromeon please opine. The speedup on MinGW/Linux is quite large, a fresh compile is at minimum 2x faster on my machine. A few points:

- PCH builds are disabled by default. Can be enabled with the `SFML_ENABLE_PCH` option.

- PCH builds are only supported on Windows/Linux. MacOS is not working for some weird reason and I cannot debug it easily.

- PCH builds are only supported for static builds of SFML. Shared library support is not worth it because the PCH cannot be reused due to mismatching macro definitions of `SFML_XXX_EXPORTS` between targets.

- It would be nice to have "static + PCH" CI jobs for Linux, MinGW, and MSVC.