@ChrisThrasher 2022-10-01
Here are my results testing this PR. I'm using an Ubuntu 22 VM running in macOS with 6 CPUs and 16 GB of memory.

---

`master` (939ec267dfa98e97438a5b76f79317c7e0738a92)
```
cmake -B build -DSFML_BUILD_EXAMPLES=ON -DSFML_BUILD_TEST_SUITE=ON -DBUILD_SHARED_LIBS=OFF
time cmake --build build -j6
```

```
Total Test time (real) =   0.35 sec
[100%] Built target runtests

real	0m13.779s
user	1m1.117s
sys	0m6.092s
```

---

`feature/use-pch` (ec0a03610d22d55fe0ed4916bf0c1dbb9076352e)
```
cmake -B build -DSFML_BUILD_EXAMPLES=ON -DSFML_BUILD_TEST_SUITE=ON -DBUILD_SHARED_LIBS=OFF -DSFML_ENABLE_PCH=ON
time cmake --build build -j6
```

```
Total Test time (real) =   0.35 sec
[100%] Built target runtests

real    0m12.228s
user    0m46.263s
sys    0m7.278s
```