@vittorioromeo 2022-10-01
My results with MSYS2/MinGW on my Windows 10 x64 laptop:

```
cmake .. -GNinja -DCMAKE_C_COMPILER=clang -DCMAKE_CXX_COMPILER=clang++ -DCMAKE_CXX_FLAGS="-ftime-trace" -DSFML_BUILD_TEST_SUITE=1 -DSFML_BUILD_EXAMPLES=1 -DSFML_BUILD_DOC=0 -DBUILD_SHARED_LIBS=0 -DSFML_ENABLE_PCH=0
```

```
time ninja
ninja  0.01s user 0.00s system 0% cpu 31.183 total
```

---

```
cmake .. -GNinja -DCMAKE_C_COMPILER=clang -DCMAKE_CXX_COMPILER=clang++ -DCMAKE_CXX_FLAGS="-ftime-trace" -DSFML_BUILD_TEST_SUITE=1 -DSFML_BUILD_EXAMPLES=1 -DSFML_BUILD_DOC=0 -DBUILD_SHARED_LIBS=0 -DSFML_ENABLE_PCH=1
```

```
time ninja
ninja  0.00s user 0.01s system 0% cpu 22.085 total
```