@vittorioromeo 2022-10-06
> This is a very nice addition to SFML, thanks a lot for this PR! 👍

Thanks!

> I'm not sure if we should switch CI over to use PCH=on; probably a lot of SFML users would still use the classic setup, and I think it's important we have the confidence that the basic configuration is not broken.

Honestly, this feature is mostly to speed up builds for people working on SFML's codebase itself. Even if it's never consumed by our users, it's still beneficial for us to have as part of our development cycle, and potentially to speed up CI.

My personal preference would be to enable PCH=on on all our CI except one job that tests a non-PCH build just to make sure it still works properly -- one of the reasons to have PCH is to speed up the whole CI process.

Happy to hear what others think.