@Bromeon 2022-10-06
> My personal preference would be to enable PCH=on on all our CI except one job that tests a non-PCH build just to make sure it still works properly -- one of the reasons to have PCH is to speed up the whole CI process.

That sounds reasonable to me. 👍