@vittorioromeo 2022-10-07
> I assume miscompilations due to PCH are no longer the dreadful thing they were 1-2 decades ago? 😄

My personal experience with PCH has been very positive in this aspect, I never had any miscompilation issue. Also, as you can see by the pain I went through, any sort of inconsistency between files causes a total compilation failure -- PCH needs every flag to be exactly the same in order to avoid weird issues. I think we should be pretty safe!