@vittorioromeo 2022-10-07
I would suggest merging this PR as-is, then having a separate PR to enable PCH in CI.