@vittorioromeo 2022-10-07
@ChrisThrasher 

> Can you share some stats for how this changes total CI runtime? Can we expect the pipeline to complete any quicker than before?

I do not have any CI stats at the moment -- the goal of this PR is to add the PCH mechanism to SFML so that we can see how much of an improvement it is. I do expect a significant speedup for all Windows and Linux static library CI builds.

I have addressed your suggestions (thanks!) -- my suggestion would be to merge this PR, create a PR to enable PCH builds where appropriate in CI, and get some rough measurements. If there's no benefit, I'd be happy to revert this PR.