@vittorioromeo 2022-10-07
SFMLBuildMaster: Build this please.